+++ code/engine/VM.cpp
@@ -247,13 +247,13 @@ LPVOID ENGINE_API neon_luamem(LPVOID ud, LPVOID ptr, size_t osize, size_t nsize)
     (void)ud;  /* not used */
 
     if (nsize == 0) {
-		gMemUsedLua -= osize;
+		gMemUsedLua -= (DWORD)osize;
         free(ptr);
         return NULL;
     }
 	else
 	{
-		gMemUsedLua += (nsize - osize);
+		gMemUsedLua += (DWORD)(nsize - osize);
 
 		neon_mempeak_update();
 		return realloc(ptr, nsize);
